Fix Authentication configuration.
Support branches with slashes.

There's an error when you add authentication configuration in the configuration. This will fix it.
When you have branches' names with slashes, i.e. feature/foo, when you use %b to add the branch name in the version number, it will replace the shash with an underscore.
[patch]